v1.2

+++ Migrations/AddTables_11092021.cs
@@ -60,7 +60,7 @@ namespace Migrations
                     .WithColumn("StudentId").AsInt64().ForeignKey("Student", "Id")
                     .WithColumn("Title").AsString()
                     .WithColumn("IsCompleted").AsBoolean()
-                    .WithColumn("Progress").AsFloat()
+                    .WithColumn("Progress").AsDouble()
                     .WithColumn("Grade").AsFloat()
                     .WithColumn("SertificateUrl").AsString().Nullable()
                     .WithColumn("LearningHours").AsFloat()
